+++ CHANGELOG.md
@@ -84,6 +84,8 @@ So your change should look like:
 - Fixed `finish()` being able to complete an already finished tween, and
   prevented `timeLeft` from going negative and `currentTime` from exceeding the
   duration (#1117) - @imaginarny
+- Fixed triangulate by updating the convexity of nearby vertices after removing
+  a concave vertex during ear cutting (#1134) - @mflerackers
 
 ## [4000.0.0-alpha.27.1] - 2026-05-12